requirements.txt
numpy
pandas
matplotlib
streamlit
jupyterlab
pyyaml